Исправлены замечания по лабам 1. Исправлен ввод файла 2. Убрано += из цикла (из-за проблем с производительностью) 3. Убраны лишние обработики ошибок 4. Добавлено возвращение объекта из функции племени 5. Добавлены функции для обработки этих объектов и возвращении нужного строкового результата
