+++ alexander.yakunin/lab4/src/UI/Menus.java
@@ -13,7 +13,6 @@ public class Menus {
         System.out.println("4. Выход");
 
         return in.nextInt();
-
     }
